feat: extract tower-mcp-types crate for lightweight MCP type access
Move protocol.rs and error.rs into a new tower-mcp-types workspace
member with minimal dependencies (serde, serde_json, thiserror, base64).
Replace originals with thin re-export wrappers so all existing
crate::protocol::* and crate::error::* paths resolve unchanged.

This enables consumers like editors (e.g. Zed) to use MCP types without
pulling in tower/tokio/axum. Motivated by rmcp#449 discussion.
