+++ Cargo.toml
@@ -14,7 +14,7 @@ keywords = ["mcp", "tower", "json-rpc", "ai", "llm"]
 categories = ["network-programming", "asynchronous"]
 
 [workspace]
-members = ["examples/conformance-server", "examples/markdownlint-mcp", "examples/codegen-mcp", "examples/notes-mcp"]
+members = ["tower-mcp-types", "examples/conformance-server", "examples/markdownlint-mcp", "examples/codegen-mcp", "examples/notes-mcp"]
 
 [workspace.dependencies]
 tower-resilience = { version = "0.7", default-features = false, features = ["bulkhead", "ratelimiter"] }
@@ -29,15 +29,15 @@ tokio = { version = "1", features = ["full"] }
 futures = "0.3"
 pin-project-lite = "0.2"
 
+# MCP protocol types
+tower-mcp-types = { version = "0.1", path = "tower-mcp-types" }
+
 # Serialization
 serde = { version = "1.0", features = ["derive"] }
 serde_json = "1.0"
 
 # Schema generation
 schemars = "1.2"
-
-# Error handling
-thiserror = "2"
 
 # Async traits
 async-trait = "0.1"
